+++ Changelog.md
@@ -79,6 +79,7 @@
 - Fix: The alarm can't play sounds added by resource packs
 - Fix: Changes to options of some customizable blocks don't get saved when leaving and re-joining a world
 - Fix: Cameras don't work when Embeddium/Rubidium is installed
+- Fix: Numerous reinforced blocks can be washed away by water
 - Removed: Ability to change the subpage in the SecurityCraft Manual using the arrow keys. This is still possible by scrolling while holding down CTRL
 - Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
 - Misc.: ProjectE support has been updated with previously missing EMC values